refactor: core/ 패키지 신설 — URL 파싱·범위 계산 이주 (#50)
- core/ 골격 생성 (api, models, downloaders, utils) + PySide6 금지 규칙 docstring 명시
- extract_content_no -> core/api/url_parser.py 이동 (코드 무변경)
- decide_part_size, split_ranges -> core/downloaders/ranges.py 이동 (코드 무변경)
- 기존 위치에 re-export/위임 유지로 호출부 무변경
- 테스트 import 경로 갱신 (기대값 무변경)
