+++ content/network.py
@@ -6,6 +6,8 @@ from http.cookiejar import DefaultCookiePolicy
 from urllib.parse import urljoin
 import xml.etree.ElementTree as ET
 
+from core.api.url_parser import extract_content_no
+
 NAVER_API = "https://apis.naver.com"
 CHZZK_API = "https://api.chzzk.naver.com"
 VIDEOHUB_API = "https://api-videohub.naver.com"
@@ -51,9 +53,8 @@ class NetworkManager:
         """
         치지직 VOD URL에서 type과 content_no를 추출한다.
 
-        브라우저 주소창에서 그대로 복사한 URL 변형을 허용한다 (#33):
-        쿼리스트링(`?t=123` 등)은 무시하고, 후행 슬래시와 www 서브도메인을 허용한다.
-        live URL·다른 도메인·형식이 깨진 URL은 계속 거부한다.
+        구현은 core/api/url_parser.py로 이동했다 (#50).
+        기존 호출부 호환을 위해 시그니처를 유지하고 core 함수에 위임한다.
 
         Args:
             vod_url (str): 치지직 VOD URL
@@ -61,16 +62,7 @@ class NetworkManager:
         Returns:
             tuple[str, str]: (type, content_no) 형식의 튜플. 매칭되지 않으면 (None, None) 반환
         """
-        if not vod_url.startswith("http://") and not vod_url.startswith("https://"):
-            vod_url = "https://" + vod_url
-        match = re.fullmatch(
-            r'https?://(?:www\.)?chzzk\.naver\.com'
-            r'/(?P<content_type>video|clips)/(?P<content_no>\w+)/?(?:\?.*)?',
-            vod_url,
-        )
-        if match:
-            return match.group("content_type"), match.group("content_no")
-        return None, None
+        return extract_content_no(vod_url)
 
     @staticmethod
     def get_video_info(video_no: str, cookies: dict):
